requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
h11==0.16.0
httptools==0.7.1
idna==3.11
librt==0.7.8
mypy==1.19.1
mypy-extensions==1.1.0
mysql-connector-python==9.5.0
passlib==1.7.4
pathspec==1.0.3
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pyjwt==2.10.1
pymysql==1.1.2
python-dotenv==1.2.1
python-jose==3.5.0
pyyaml==6.0.3
rsa==4.9.1
six==1.17.0
sqlalchemy==2.0.46
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
